added second data collection and query counts
Today I tested the performance of our queries within the first data collection. I found out that the singular form of a word does not return tweets in which the word is in its plural form. For this reason I have collected data on the plural form of the word separately. The data here added comprises:

-  second_dataset_chatgpt.json is the json output of searching:
and <- "chatgpt"
or <- c("employees", # post collection insertion
        "jobs", # post collection insertion
        "skills", # post collection insertion
        "concerns", # post collection insertion
        "threats", # post collection insertion
        "occupations", # post collection insertion
        "earnings", # post collection insertion
        "tasks", # post collection insertion
        "collaborations") # post collection insertion

The collection returned just over 93 thousand tweets. 

- query_counts.csv is a csv that gives us information on the number of tweets resulting from all our queries, plural and singular forms of a word both included. It can be used to compare differences in performance of keywords in terms of number of tweets retrieved.